requirements.txt
streamlit>=1.28.0
ultralytics>=8.0.0
pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0